backend/requirements.txt
boto==2.49.0
boto3==1.9.76
botocore==1.12.76
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-storages-redux==1.3.3
djangorestframework==3.9.0
docutils==0.14
jmespath==0.9.3
Markdown==3.0.1
mysqlclient==1.3.14
Pillow==5.4.1
python-dateutil==2.7.5
pytz==2018.9
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
